nvofapi: Refactor to match suggested layout from Jens
